Create  CRASH A WEBSITE!💩
I am not responsible if this causes your browser, webpage, or computer to crash, and I hope you use this cautiously. This is not supposed to be helpful or life-changing, but if you are naughty try it. If you wanna use it enter it in Console (Right-click then hit console.)